Task 5: extract shared normalizeChildren utility
Duplicate implementations existed in src/parser/index.ts and
src/transformers/layout/container.ts. Moved to
src/utils/normalize-children.ts and imported in both places.
