+++ src/transformers/layout/container.ts
@@ -1,30 +1,9 @@
 import {type Props as ContainerProperties} from '../../components/layout/container';
 import {type Element, type Child} from '../../constants/types';
 import {transform} from '..';
+import normalizeChildren from '../../utils/normalize-children';
 
 type ContainerType = Child[];
-
-const normalizeChildren = (children: Child): Child[] => {
-  const result: Child[] = [];
-  const stack = Array.isArray(children) ? [...children] : [children];
-
-  while (stack.length > 0) {
-    const child = stack.shift();
-
-    if (child === null || child === undefined || typeof child === 'boolean') {
-      continue;
-    }
-
-    if (Array.isArray(child)) {
-      stack.unshift(...child);
-      continue;
-    }
-
-    result.push(child);
-  }
-
-  return result;
-};
 
 const transformContainer = (child: Element): ContainerType => {
   const {children}: ContainerProperties = child.props;
